+++ structs/CollegePlayerProgressions.go
@@ -1,17 +1,17 @@
 package structs
 
 type CollegePlayerProgressions struct {
-	Age              int
-	Stars            int
-	Overall          int
-	Shooting2        int
-	Shooting3        int
-	Finishing        int
-	Ballwork         int
-	Rebounding       int
-	InteriorDefense  int
-	PerimeterDefense int
-	FreeThrow        int
-	Stamina          int
+	Age              uint8
+	Stars            uint8
+	Overall          uint8
+	Shooting2        uint8
+	Shooting3        uint8
+	Finishing        uint8
+	Ballwork         uint8
+	Rebounding       uint8
+	InteriorDefense  uint8
+	PerimeterDefense uint8
+	FreeThrow        uint8
+	Stamina          uint8
 	PotentialGrade   string
 }
